refactor(quick-filters): extract checkbox value fetching into hook
